payment/pyproject.toml
[project]
name = "payment"
version = "0.1.0"
description = "payment service"
authors = [
    {name = "AMKalinin",email = "am1kalinin@gmail.com"}
]
readme = "README.md"
requires-python = "3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "python-keycloak (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
